docker-images/apis/rest-memory/requirements.txt
py-easy-rest==0.4.0
PyYAML==5.4.1